NOT-63 cleaned up edit note component
