+++ src/lib/components/EditNote/EditNote.svelte
@@ -9,22 +9,16 @@
 		ImageEditor,
 		FormGroup,
 		ValidationErrors,
-		Cross,
 		Plus,
 	} from '$lib/components'
 	import type { EditNoteProps } from './types.editNote'
 	import { NoteEditorSchema } from '$lib/schemas'
-	import { generateCopy } from './editNote.helpers'
+	import { createState, generateCopy } from './helpers.editNote.svelte'
 	import { idTestId, imageListTestId } from './consts.editNote'
 	import InfoBar from './InfoBar.svelte'
 
 	let { data, action, images }: EditNoteProps = $props()
-	let imageList = $state(
-		images?.length
-			? images
-			: [{ id: undefined, file: undefined, altText: undefined }],
-	)
-
+	const helperState = createState(images)
 	const { form, errors, enhance, formId, constraints, delayed, timeout } =
 		superForm(data, {
 			validators: zodClient(NoteEditorSchema),
@@ -33,13 +27,6 @@
 		$form.id,
 		$form.title,
 	)
-
-	function addEmptyImage() {
-		imageList = [
-			...imageList,
-			{ id: undefined, file: undefined, altText: undefined },
-		]
-	}
 </script>
 
 <form
@@ -54,7 +41,6 @@
 	{#if $form.id}
 		<input hidden value={$form.id} name="id" data-testid={idTestId} />
 	{/if}
-	<!-- TODO: Focus first input -->
 	<FormGroup flex="0">
 		<Input
 			errors={$errors.title}
@@ -64,6 +50,7 @@
 			type="text"
 			bind:value={$form.title}
 			constraints={$constraints.title}
+			autofocus
 		/>
 	</FormGroup>
 	<FormGroup flex="1 0 0">
@@ -80,29 +67,11 @@
 	</FormGroup>
 	<span>Images</span>
 	<ul class="image-list" data-testid={imageListTestId}>
-		{#each imageList as image, index (index)}
-			<li class="image-list-item">
-				<!-- TODO: //create form action for delete later NOT-46 -->
-				<button
-					formaction="?/delete"
-					class="remove-image-button"
-					name="id"
-					value={image.id ?? index}
-					onclick={() => {
-						imageList = imageList.filter((_, i) => i !== index)
-					}}
-				>
-					<span aria-hidden="true">
-						<Cross />
-					</span>
-					<span class="sr-only">Remove image {index}</span>
-				</button>
-				<ImageEditor {image} {index} />
-			</li>
+		{#each helperState.imageList as image, index (index)}
+			<ImageEditor {image} {index} deleteImage={helperState.deleteImage} />
 		{/each}
 	</ul>
-	<!-- TODO: AddEmptyImage is not working 🐛 NOT-65 -->
-	<Button variant="secondary" type="button" onclick={addEmptyImage}>
+	<Button variant="secondary" type="button" onclick={helperState.addEmptyImage}>
 		<Plus />
 		Add another image
 	</Button>
@@ -138,26 +107,5 @@
 		list-style: none;
 		display: flex;
 		flex-direction: column;
-	}
-
-	.image-list-item {
-		position: relative;
-		padding-bottom: var(--space-xs);
-	}
-
-	.remove-image-button {
-		display: grid;
-		place-items: center;
-		position: absolute;
-		right: 0;
-		top: 0;
-		border: none;
-		background: none;
-		color: tomato;
-		cursor: pointer;
-	}
-
-	.remove-image-button span {
-		display: flex;
 	}
 </style>
