+++ src/lib/components/ImageEditor/types.imageEditor.ts
@@ -3,4 +3,5 @@ import type { ImageFieldset } from '../EditNote/types.editNote'
 export type ImageEditorProps = {
 	image: ImageFieldset | undefined
 	index: number
+	deleteImage: (index: number) => void
 }
